Crate arcdps_parse

Parsing for ArcDPS EVTC logs.

Usage

Use the parse_file function to easily parse a Log from a file path.

match arcdps_parse::parse_file("path/to/log.evtc") {
    Ok(log) => println!("Log for boss id {}", log.header.boss_id),
    Err(err) => eprintln!("Encountered error {}", err),
}

A Log can also be parsed from any input implementing Read.

use arcdps_parse::{Log, Parse};
use std::io;

fn parse_from_read(input: &mut impl io::Read) -> Log {
    Log::parse(input).expect("failed to parse")
}

Note that ArcDPS can save compressed log files with .zevtc as file extension. Enabling the "zevtc" or "zip" feature adds support for compressed logs.

Structs

ActivationIter

An iterator over the variants of Self

AffinityIter

An iterator over the variants of Self

Agent

An EVTC agent.

AgentKindIter

An iterator over the variants of Self

AttributeIter

An iterator over the variants of Self

BuffCycleIter

An iterator over the variants of Self

BuffFormula

Buff formula from a CombatEvent with StateChange::BuffFormula.

BuffInfo

Buff information from a CombatEvent with StateChange::BuffInfo.

BuffRemoveIter

An iterator over the variants of Self

CombatEvent

ArcDPS combat event.

ContentLocalIter

An iterator over the variants of Self

Effect

Effect information from a CombatEvent with StateChange::Effect.

EventKindIter

An iterator over the variants of Self

Header

An EVTC log header.

LanguageIter

An iterator over the variants of Self

Log

An EVTC log.

Position

Positional information.

ProfessionIter

An iterator over the variants of Self

RawCombatEvent

Skill

An EVTC skill definition.

SkillInfo

Skill information from a CombatEvent with StateChange::SkillInfo.

SkillTiming

Skill timing from a CombatEvent with StateChange::SkillTiming.

SpecializationIter

An iterator over the variants of Self

StateChangeIter

An iterator over the variants of Self

StrikeIter

An iterator over the variants of Self

WeaponSetIter

An iterator over the variants of Self

Enums

Activation

Skill activation (cast).

Affinity

Whether the agent is an ally or enemy.

AgentKind

Possible [Agent] kinds.

Attribute

Attributes for buff formulas.

BuffCategory

Buff info category.

BuffCycle

Combat buff cycle.

BuffRemove

Combat buff remove.

ContentLocal

Content local for StateChange::IdToGUID events.

CustomSkill

ArcDPS custom skill ids.

EffectDuration

Duration of an effect in time or as a tracking id.

EventKind

Possible CombatEvent kinds.

Language

GW2 client language.

ParseError

A possible error occurring during parsing.

Profession

Player profession.

Specialization

Player specialization.

StateChange

Combat state change.

Strike

Strike types.

WeaponSet

Agent weapon set.

Traits

Parse

Interface for parsing a value from a Read input.

Save

Interface for saving a value into a Write output.

Functions

parse_file

Parses a Log from a given Path to a log file.

parse_zevtc

Parses a Log from a compressed zevtc input.