Struct arcdps_parse::CombatEvent

pub struct CombatEvent {

    pub time: u64,
    pub src_agent: u64,
    pub dst_agent: u64,
    pub value: i32,
    pub buff_dmg: i32,
    pub overstack_value: u32,
    pub skill_id: u32,
    pub src_instance_id: u16,
    pub dst_instance_id: u16,
    pub src_master_instance_id: u16,
    pub dst_master_instance_id: u16,
    pub affinity: Affinity,
    pub buff: u8,
    pub result: u8,
    pub is_activation: Activation,
    pub is_buff_remove: BuffRemove,
    pub is_ninety: u8,
    pub is_fifty: u8,
    pub is_moving: u8,
    pub is_statechange: StateChange,
    pub is_flanking: u8,
    pub is_shields: u8,
    pub is_off_cycle: u8,
    pub pad61: u8,
    pub pad62: u8,
    pub pad63: u8,
    pub pad64: u8,

}

ArcDPS combat event.

Fields
time: u64

timeGetTime() at time of registering the event.

src_agent: u64

Agent that caused the event.

dst_agent: u64

Agent the event happened to.

value: i32

Value, if relevant to the event.

buff_dmg: i32

Buff damage, if relevant to the event.

overstack_value: u32

Overstack value, if relevant to the event.

skill_id: u32

Skill id of the relevant skill (can be zero).

src_instance_id: u16

Instance id of source agent as appears in game at time of event.

dst_instance_id: u16

Instance id of destination agent as appears in game at time of event.

src_master_instance_id: u16

If src_agent has a master (e.g. is minion), will be equal to instance id of master, zero otherwise.

dst_master_instance_id: u16

If dst_agent has a master (e.g. is minion), will be equal to instance id of master, zero otherwise.

affinity: Affinity

Current affinity of src_agent and dst_agent.

Arc calls this “iff” for if friend/foe.

buff: u8

Buff, if relevant to the event.

result: u8

Combat result.

For direct damage events this is a variant of Strike.

is_activation: Activation

Whether event is a kind of Activation.

For Activation::CancelFire and Activation::CancelCancel value will be the ms duration of the time spent in animation. buff_dmg will be the ms duration of the scaled (as if not affected) time spent.

For Normal or Quickness, value will be the ms duration at which all significant effects have happened. buff_dmg will be the ms duration at which control is expected to be returned to character.

dst_agent will be x/y of target of skill effect. overstack_value will be z of target of skill effect.

is_buff_remove: BuffRemove

Whether event is a kind of BuffRemove.

src_agent is agent that had buff removed, dst_agent is the agent that removed it. value will be the remaining time removed calculated as duration. buff_dmg will be the remaining time removed calculated as intensity.

For BuffRemove::All result will be the number of stacks removed. For BuffRemove::Single pad61-64 (uint32) will be buff instance id of buff removed.

is_ninety: u8

Whether src_agent is above 90% Health.

is_fifty: u8

Whether dst_agent is below 50% Health.

is_moving: u8

Whether src_agent is moving at time of event.

is_statechange: StateChange

Whether event is a kind of StateChange.

is_flanking: u8

Whether src_agent is flanking at time of event.

The value lies in a range of 1 to 135 degrees where 135 is rear.

is_shields: u8

Shields, if relevant to the event.

is_off_cycle: u8

For relevant events this may be a variant of BuffCycle.

pad61: u8

Padding.

May contain information depending on the kind of event.

pad62: u8

Padding.

May contain information depending on the kind of event.

pad63: u8

Padding.

May contain information depending on the kind of event.

pad64: u8

Padding.

May contain information depending on the kind of event.

Implementations

impl CombatEvent

pub fn kind(&self) -> EventKind

Determines the kind of event.

pub fn has_time(&self) -> bool

Checks whether the event has a timestamp.

pub fn time(&self) -> Option<u64>

Retrieves the event time, if present.

pub fn position(&self) -> Option<Position>

Attempts to extract Position data from the event.

pub fn effect(&self) -> Option<Effect>

Attempts to extract Effect data from the event.

pub fn effect_old(&self) -> Option<EffectOld>

Attempts to extract EffectOld data from the event.

pub fn effect_guid(&self) -> Option<EffectGUID>

Attempts to extract EffectGUID information from the event.

pub fn buff_info(&self) -> Option<BuffInfo>

Attempts to extract BuffInfo from the event.

pub fn raw_buff_formula(&self) -> Option<RawBuffFormula>

Attempts to extract a RawBuffFormula from the event.

pub fn buff_formula(&self) -> Option<BuffFormula>

Attempts to extract a BuffFormula from the event.

pub fn skill_info(&self) -> Option<SkillInfo>

Attempts to extract SkillInfo from the event.

pub fn skill_timing(&self) -> Option<SkillTiming>

Attempts to extract SkillTiming data from the event.