Trait arcdps_parse::Parse

pub trait Parse: Sized {
    type Error;

    // Required method
    fn parse(input: &mut impl Read) -> Result<Self, Self::Error>;

    // Provided method
    fn parse_multi(
        input: &mut impl Read,
        count: usize
    ) -> Result<Vec<Self>, Self::Error> { ... }
}

Interface for parsing a value from a Read input.

Required Associated Types

type Error

Associated error which can happen during parsing.

Required Methods

fn parse(input: &mut impl Read) -> Result<Self, Self::Error>

Parses a value of this type from the input.

Provided Methods

fn parse_multi( input: &mut impl Read, count: usize ) -> Result<Vec<Self>, Self::Error>

Parses multiple values of this type from the input into a Vec.