Crate arcdps_parse

Parsing for ArcDPS EVTC logs.

Usage

A Log can be parsed from any input implementing Read.

use arcdps_parse::{Log, Parse};
use std::fs::File;

let mut file = File::open("log.evtc")?;
let log = Log::parse(&mut file)?;
assert_eq!(log.header.revision, 1);

Note that ArcDPS may save zipped log files with the .zevtc extension. Reading those can be realized using for example the zip crate.

Structs

AffinityIter

An iterator over the variants of Self

Agent

An EVTC agent.

AgentKindIter

An iterator over the variants of Self

AttributeIter

An iterator over the variants of Self

BuffCycleIter

An iterator over the variants of Self

BuffRemoveIter

An iterator over the variants of Self

CombatEvent

ArcDPS combat event.

ContentLocalIter

An iterator over the variants of Self

CustomSkillIter

An iterator over the variants of Self

EventKindIter

An iterator over the variants of Self

Header

An EVTC log header.

LanguageIter

An iterator over the variants of Self

Log

An EVTC log.

ProfessionIter

An iterator over the variants of Self

RawCombatEvent

Skill

An EVTC skill definition.

SpecializationIter

An iterator over the variants of Self

StateChangeIter

An iterator over the variants of Self

StrikeIter

An iterator over the variants of Self

Enums

Activation

Skill activation (cast).

Affinity

Whether the agent is an ally or enemy.

AgentKind

Possible [Agent] kinds.

Attribute

Attributes for buff formulas.

BuffCategory

Buff info category.

BuffCycle

Combat buff cycle.

BuffRemove

Combat buff remove.

ContentLocal

Content local for StateChange::IdToGUID events.

CustomSkill

ArcDPS custom skill ids.

EventKind

Possible CombatEvent kinds.

Language

GW2 client language.

ParseError

A possible error occurring during parsing.

Profession

Player profession.

Specialization

Player specialization.

StateChange

Combat state change.

Strike

Strike types.

Traits

Parse

Interface for parsing a value from a Read input.