Crate arcdps_parse

Parsing for ArcDPS EVTC logs.

Usage

Use the parse_file function to easily parse a Log from a file path.

match arcdps_parse::parse_file("path/to/log.evtc") {
    Ok(log) => println!("Log for boss id {}", log.header.boss_id),
    Err(err) => eprintln!("Encountered error {}", err),
}

A Log can also be parsed from any input implementing Read.

use arcdps_parse::{Log, Parse};
use std::io;

fn parse_from_read(input: &mut impl io::Read) -> Log {
    Log::parse(input).expect("failed to parse")
}

Note that ArcDPS can save compressed log files with .zevtc as file extension. Enabling the "zevtc" or "zip" feature adds support for compressed logs.

Structs

ActivationIter

An iterator over the variants of Activation

AffinityIter

An iterator over the variants of Affinity

Agent

An EVTC agent.

AgentKindIter

An iterator over the variants of AgentKind

AttributeIter

An iterator over the variants of Attribute

BreakbarStateIter

An iterator over the variants of BreakbarState

BuffCycleIter

An iterator over the variants of BuffCycle

BuffFormula

Buff formula from a CombatEvent with StateChange::BuffFormula.

BuffInfo

Buff information from a CombatEvent with StateChange::BuffInfo.

BuffRemoveIter

An iterator over the variants of BuffRemove

BuffStackTypeIter

An iterator over the variants of BuffStackType

CombatEvent

ArcDPS combat event.

ConditionResultIter

An iterator over the variants of ConditionResult

ContentLocalIter

An iterator over the variants of ContentLocal

Effect

Effect information from a CombatEvent with StateChange::Effect.

EffectGUID

Effect information from a CombatEvent with StateChange::IdToGUID.

EffectOld

Effect information from a CombatEvent with StateChange::EffectOld.

EffectOrientation

Orientation of an effect.

EventKindIter

An iterator over the variants of EventKind

Header

An EVTC log header.

LanguageIter

An iterator over the variants of Language

Log

An EVTC log.

Position

Positional information.

ProfessionIter

An iterator over the variants of Profession

RawBuffFormula

Buff formula from a CombatEvent with StateChange::BuffFormula.

RawCombatEvent

Skill

An EVTC skill definition.

SkillInfo

Skill information from a CombatEvent with StateChange::SkillInfo.

SkillTiming

Skill timing from a CombatEvent with StateChange::SkillTiming.

SpecializationIter

An iterator over the variants of Specialization

StateChangeIter

An iterator over the variants of StateChange

StrikeIter

An iterator over the variants of Strike

WeaponSetIter

An iterator over the variants of WeaponSet

Enums

Activation

Skill activation (cast).

Affinity

Whether the agent is an ally or enemy.

AgentKind

Possible agent kinds.

Attribute

Attributes for buff formulas.

BreakbarState

Breakbar (defiance bar) states.

BuffCategory

Buff info category after 13 December 2022.

BuffCategoryOld

Buff info category before 13 December 2022.

BuffCycle

Combat buff cycle.

BuffRemove

Combat buff remove.

BuffStackType

Buff stacking behavior.

ConditionResult

Condition tick results.

ContentLocal

Content local for StateChange::IdToGUID events.

CustomSkill

ArcDPS custom skill ids.

EffectDuration

Duration of an effect in time or as a tracking id.

EffectLocation

Location of an effect.

EventKind

Possible CombatEvent kinds.

Language

GW2 client language.

ParseError

A possible error occurring during parsing.

Profession

Player profession.

Specialization

Player specialization.

StateChange

Combat state change.

Strike

Strike types.

WeaponSet

Agent weapon set.

Traits

Parse

Interface for parsing a value from a Read input.

Save

Interface for saving a value into a Write output.

Functions

parse_file

Parses a Log from a given Path to a log file.

parse_zevtc

Parses a Log from a compressed zevtc input.